refactor: remove termynal terminal emulator and reorganize insights directory structure
