+++ pyproject.toml
@@ -28,7 +28,6 @@ mkdocs = "^1.6.1"
 mkdocs-material = "^9.5.39"
 mkdocstrings = { extras = ["python"], version = "^0.29.0" }
 mkdocs-autorefs = "^1.4.1"
-termynal = "^0.12.1"
 pytest = "^8.3.3"
 factory-boy = "^3.3.1"
 invoke = "^2.2.0"
